requirements.txt
fastapi
uvicorn
pyjwt
python
redis
httpx
prometheus_client
jinja2
python_jose
openapi_core
PyYAML
opentelemetry_api
opentelemetry_sdk
opentelemetry_exporter_otlp_proto_http
opentelemetry_instrumentation_fastapi
opentelemetry_instrumentation_logging
python_multipart
pydantic_settings